move CryptographicHydrator into Cryptography namespace
